requirements.txt
pyTelegramBotAPI==3.8.1
PySocks==1.7.1